ml/perspective-classification/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
tqdm
matplotlib
scikit-learn
